Move aggregation functions to separate module
